added -Wextra to compile flags, fixed corresponding warnings

+++ sopnet/features/SetDifference.cpp
@@ -8,15 +8,15 @@ SetDifference::operator()(const Slice& slice1, const Slice& slice2, bool normali
 
 	const util::rect<double>& bb = slice1.getComponent()->getBoundingBox();
 
-	util::point<unsigned int> offset(static_cast<unsigned int>(bb.minX), static_cast<unsigned int>(bb.minY));
-	util::point<unsigned int> size(static_cast<unsigned int>(bb.width() + 2), static_cast<unsigned int>(bb.height() + 2));
+	util::point<int> offset(static_cast<int>(bb.minX), static_cast<int>(bb.minY));
+	util::point<int> size(static_cast<int>(bb.width() + 2), static_cast<int>(bb.height() + 2));
 
 	std::vector<bool> pixels(size.x*size.y, false);
 
 	foreach (const util::point<unsigned int>& pixel, slice1.getComponent()->getPixels()) {
 
-		unsigned int x = pixel.x - offset.x;
-		unsigned int y = pixel.y - offset.y;
+		int x = (int)pixel.x - offset.x;
+		int y = (int)pixel.y - offset.y;
 
 		pixels[x + y*size.x] = true;
 	}
@@ -47,22 +47,22 @@ SetDifference::operator()(const Slice& slice1a, const Slice& slice1b, const Slic
 
 	util::rect<double> bb(std::min(bba.minX, bbb.minX), std::min(bba.minY, bbb.minY), std::max(bba.maxX, bbb.maxX), std::max(bba.maxY, bbb.maxY));
 
-	util::point<unsigned int> offset(static_cast<unsigned int>(bb.minX), static_cast<unsigned int>(bb.minY));
-	util::point<unsigned int> size(static_cast<unsigned int>(bb.width() + 2), static_cast<unsigned int>(bb.height() + 2));
+	util::point<int> offset(static_cast<int>(bb.minX), static_cast<int>(bb.minY));
+	util::point<int> size(static_cast<int>(bb.width() + 2), static_cast<int>(bb.height() + 2));
 
 	std::vector<bool> pixels(size.x*size.y, false);
 
 	foreach (const util::point<unsigned int>& pixel, slice1a.getComponent()->getPixels()) {
 
-		unsigned int x = pixel.x - offset.x;
-		unsigned int y = pixel.y - offset.y;
+		int x = (int)pixel.x - offset.x;
+		int y = (int)pixel.y - offset.y;
 
 		pixels[x + y*size.x] = true;
 	}
 	foreach (const util::point<unsigned int>& pixel, slice1b.getComponent()->getPixels()) {
 
-		unsigned int x = pixel.x - offset.x;
-		unsigned int y = pixel.y - offset.y;
+		int x = (int)pixel.x - offset.x;
+		int y = (int)pixel.y - offset.y;
 
 		pixels[x + y*size.x] = true;
 	}
@@ -96,7 +96,7 @@ unsigned int
 SetDifference::numDifferent(
 		unsigned int size1,
 		const std::vector<bool>& pixels,
-		const util::point<unsigned int>& size,
+		const util::point<int>& size,
 		const util::point<double>& centerOffset,
 		const util::point<double>& offset,
 		const Slice& slice2) {
@@ -109,8 +109,8 @@ SetDifference::numDifferent(
 
 	foreach (const util::point<unsigned int>& pixel, slice2.getComponent()->getPixels()) {
 
-		unsigned int x = pixel.x - centerOffset.x - offset.x;
-		unsigned int y = pixel.y - centerOffset.x - offset.y;
+		int x = (int)pixel.x - centerOffset.x - offset.x;
+		int y = (int)pixel.y - centerOffset.x - offset.y;
 
 		// not even close
 		if (x < 0 || x >= size.x || y < 0 || y >= size.y) {
